refactor(kitchen pos) : service -> application 잘못된 패키지 이동
