+++ src/main/java/kitchenpos/order/eatinorder/ui/EatInOrderRestController.java
@@ -3,8 +3,8 @@ package kitchenpos.order.eatinorder.ui;
 import java.net.URI;
 import java.util.List;
 import java.util.UUID;
-import kitchenpos.order.eatinorder.service.EatInOrderService;
-import kitchenpos.order.eatinorder.service.dto.EatInOrderServiceRs;
+import kitchenpos.order.eatinorder.application.EatInOrderService;
+import kitchenpos.order.eatinorder.application.dto.EatInOrderServiceRs;
 import kitchenpos.order.eatinorder.ui.dto.CreateEatInOrderRq;
 import kitchenpos.order.eatinorder.ui.dto.EatInOrderRs;
 import org.springframework.http.ResponseEntity;
