+++ src/main/java/kitchenpos/order/eatinorder/ui/OrderTableRestController.java
@@ -5,8 +5,8 @@ import java.util.List;
 import java.util.UUID;
 import kitchenpos.order.eatinorder.application.OrderTableService;
 import kitchenpos.order.eatinorder.domain.model.OrderTable;
-import kitchenpos.order.eatinorder.service.dto.CreateOrderTableServiceRq;
-import kitchenpos.order.eatinorder.service.dto.OrderTableServiceRs;
+import kitchenpos.order.eatinorder.application.dto.CreateOrderTableServiceRq;
+import kitchenpos.order.eatinorder.application.dto.OrderTableServiceRs;
 import kitchenpos.order.eatinorder.ui.dto.CreateOrderTableRq;
 import kitchenpos.order.eatinorder.ui.dto.OrderTableRs;
 import org.springframework.http.ResponseEntity;
