+++ src/test/java/kitchenpos/order/eatinorder/application/OrderTableServiceTest.java
@@ -23,8 +23,8 @@ import kitchenpos.order.eatinorder.domain.repository.EatInOrderRepository;
 import kitchenpos.order.eatinorder.domain.repository.OrderTableRepository;
 import kitchenpos.order.eatinorder.domain.service.OrderTableOccupationManager;
 import kitchenpos.order.eatinorder.infra.persistence.FakeEatInOrderRepository;
-import kitchenpos.order.eatinorder.service.dto.CreateOrderTableServiceRq;
-import kitchenpos.order.eatinorder.service.dto.OrderTableServiceRs;
+import kitchenpos.order.eatinorder.application.dto.CreateOrderTableServiceRq;
+import kitchenpos.order.eatinorder.application.dto.OrderTableServiceRs;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
